drivers: misc: pn547: Consolidate driver into single folder
To comply with standards found elsewhere (copylefts, mainline nfc
drivers in drivers/nfc) move the entire PN547 driver into a subfolder of
drivers/misc/. This also moves the header file that is not shared with
any other code in the kernel and has no reason to reside under
include/linux.
