+++ drivers/misc/pn547/Makefile
@@ -1,0 +1,2 @@
+
+obj-$(CONFIG_NFC_PN547)		+= pn547.o
